Created new utility for directly monitoring file changes
gbt_config_watch: watches for file changes to GBT config area using inotifywait. If it detects changes to dirty files, it appends the diff to a log file.

gbt_config_notify: intended for use as a cronjob. When run, it examines the log file -- if it is non-empty, it is parsed and sent via email to the given recipients